Pushed all kind of message to the outboxes of the client thread an performed organization.
Idea: perhaps I should merge some kinds of message to one message buffer. Both to extend the set bounds of each message in a buffer_state_t and to organize a link to each other following or associative message/command that has a strong tie.
I.e. it would make multiple messages to be merged/entangled with another message.